+++ src/gpo_lens/__init__.py
@@ -21,4 +21,4 @@ __all__ = [
     "Som",
     "SomLink",
 ]
-__version__ = "0.6.4"
+__version__ = "0.7.0"